fix backlight

+++ kl58b.h
@@ -14,7 +14,7 @@
 
 
 extern volatile uint16_t k58b_timer;
-extern volatile uint16_t k58b_pw; // prozentwert fuer pwm
+extern volatile uint8_t k58b_pw; // prozentwert fuer pwm
 
 void initk58_pwm(void);
 void set_backlight(uint8_t pc);
